bot/requirements.txt
aiogram==3.3.0
aiohttp
telegram-wallet-pay>=1.0.0
requests
schedule
python-telegram-bot==20.4
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9